requirements.txt
# Flask and Web Framework
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
Werkzeug==3.0.1

# MongoDB Atlas (UPDATED - No more Supabase!)
pymongo==4.6.1
dnspython==2.4.2

# OpenAI
openai==1.6.1

# Twilio (WhatsApp)
twilio==8.11.0
PyJWT==2.8.0

# NLP and ML
numpy==1.26.2
scikit-learn==1.3.2
spacy==3.7.2
nltk==3.8.1
pandas==2.1.4

# Async Support
aiohttp==3.9.1

# Date/Time
python-dateutil==2.8.2

# HTTP Requests
requests==2.31.0
urllib3==2.1.0

# Production Server
gunicorn==21.2.0

# Data Validation
pydantic==2.5.3

# Additional utilities
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6

# Testing (Optional)
pytest==7.4.3
pytest-asyncio==0.21.17